remove narrowing conversion to prepare for C++11
warning: narrowing conversion of ‘’ from ‘’ to ‘’ inside { } is ill-formed in C++11 [-Wnarrowing]

+++ mythtv/libs/libmyth/programtypes.h
@@ -45,7 +45,7 @@ MPUBLIC extern const char *kJobQueueInUseID;
 MPUBLIC extern const char *kCCExtractorInUseID;
 
 /// Frame # -> File offset map
-typedef QMap<uint64_t, uint64_t> frm_pos_map_t;
+typedef QMap<long long, long long> frm_pos_map_t;
 
 typedef enum {
     MARK_ALL           = -100,
